Updates to auth, using separate module for validation checks
