+++ index.js
@@ -1,12 +1,11 @@
-/*
+/***********************************************************************************************************************************
  File: index.js
  Author(s): Russell Alexander
  Description: 
   This file is the entry point for the application. It sets up the express server and defines the routes.
  Date Created: 2024-09-26
- Last Modified: 2024-10-16
-*/
-
+ Last Modified: 2024-10-17
+***********************************************************************************************************************************/
 
 // imports
 const express = require('express');
